Move BadHttpRequestException to Http.Abstractions (#20339)
