+++ src/Http/Http.Abstractions/ref/Microsoft.AspNetCore.Http.Abstractions.netcoreapp.cs
@@ -97,6 +97,14 @@ namespace Microsoft.AspNetCore.Cors.Infrastructure
 }
 namespace Microsoft.AspNetCore.Http
 {
+    public partial class BadHttpRequestException : System.IO.IOException
+    {
+        public BadHttpRequestException(string message) { }
+        public BadHttpRequestException(string message, System.Exception innerException) { }
+        public BadHttpRequestException(string message, int statusCode) { }
+        public BadHttpRequestException(string message, int statusCode, System.Exception innerException) { }
+        public int StatusCode { [System.Runtime.CompilerServices.CompilerGeneratedAttribute] get { throw null; } }
+    }
     public abstract partial class ConnectionInfo
     {
         protected ConnectionInfo() { }
