+++ src/Servers/Kestrel/Core/test/MessageBodyTests.cs
@@ -221,7 +221,9 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
                 input.Add("g");
                 input.Add("g");
 
+#pragma warning disable CS0618 // Type or member is obsolete
                 await Assert.ThrowsAsync<BadHttpRequestException>(() => task);
+#pragma warning restore CS0618 // Type or member is obsolete
 
                 await body.StopAsync();
             }
@@ -247,7 +249,9 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
                     input.Add(i.ToString());
                 }
 
+#pragma warning disable CS0618 // Type or member is obsolete
                 await Assert.ThrowsAsync<BadHttpRequestException>(() => task);
+#pragma warning restore CS0618 // Type or member is obsolete
 
                 await body.StopAsync();
             }
@@ -286,7 +290,9 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
                 input.Add("\r");
                 input.Add("n");
 
+#pragma warning disable CS0618 // Type or member is obsolete
                 await Assert.ThrowsAsync<BadHttpRequestException>(() => task);
+#pragma warning restore CS0618 // Type or member is obsolete
 
                 await body.StopAsync();
             }
@@ -388,7 +394,9 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
                 input.Add("012345678\r");
 
                 var buffer = new byte[1024];
+#pragma warning disable CS0618 // Type or member is obsolete
                 var ex = await Assert.ThrowsAsync<BadHttpRequestException>(async () =>
+#pragma warning restore CS0618 // Type or member is obsolete
                     await stream.ReadAsync(buffer, 0, buffer.Length));
 
                 Assert.Equal(CoreStrings.BadRequest_BadChunkSizeData, ex.Message);
@@ -536,7 +544,9 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
         {
             using (var input = new TestInput())
             {
+#pragma warning disable CS0618 // Type or member is obsolete
                 var ex = Assert.Throws<BadHttpRequestException>(() =>
+#pragma warning restore CS0618 // Type or member is obsolete
                     Http1MessageBody.For(HttpVersion.Http11, new HttpRequestHeaders { HeaderTransferEncoding = "chunked, not-chunked" }, input.Http1Connection));
 
                 Assert.Equal(StatusCodes.Status400BadRequest, ex.StatusCode);
@@ -553,7 +563,9 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
             using (var input = new TestInput())
             {
                 input.Http1Connection.Method = method;
+#pragma warning disable CS0618 // Type or member is obsolete
                 var ex = Assert.Throws<BadHttpRequestException>(() =>
+#pragma warning restore CS0618 // Type or member is obsolete
                     Http1MessageBody.For(HttpVersion.Http11, new HttpRequestHeaders(), input.Http1Connection));
 
                 Assert.Equal(StatusCodes.Status411LengthRequired, ex.StatusCode);
@@ -570,7 +582,9 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
             using (var input = new TestInput())
             {
                 input.Http1Connection.Method = method;
+#pragma warning disable CS0618 // Type or member is obsolete
                 var ex = Assert.Throws<BadHttpRequestException>(() =>
+#pragma warning restore CS0618 // Type or member is obsolete
                     Http1MessageBody.For(HttpVersion.Http10, new HttpRequestHeaders(), input.Http1Connection));
 
                 Assert.Equal(StatusCodes.Status400BadRequest, ex.StatusCode);
@@ -733,7 +747,9 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
                 // Time out on the next read
                 input.Http1Connection.SendTimeoutResponse();
 
+#pragma warning disable CS0618 // Type or member is obsolete
                 var exception = await Assert.ThrowsAsync<BadHttpRequestException>(async () => await body.ReadAsync());
+#pragma warning restore CS0618 // Type or member is obsolete
                 Assert.Equal(StatusCodes.Status408RequestTimeout, exception.StatusCode);
 
                 await body.StopAsync();
@@ -768,7 +784,9 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
 
                 mockLogger.Verify(logger => logger.ConnectionBadRequest(
                     It.IsAny<string>(),
+#pragma warning disable CS0618 // Type or member is obsolete
                     It.Is<BadHttpRequestException>(ex => ex.Reason == RequestRejectionReason.RequestBodyTimeout)));
+#pragma warning restore CS0618 // Type or member is obsolete
 
                 await body.StopAsync();
             }
@@ -796,7 +814,9 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
 
                 using (var ms = new MemoryStream())
                 {
+#pragma warning disable CS0618 // Type or member is obsolete
                     var exception = await Assert.ThrowsAsync<BadHttpRequestException>(() => stream.CopyToAsync(ms));
+#pragma warning restore CS0618 // Type or member is obsolete
                     Assert.Equal(StatusCodes.Status408RequestTimeout, exception.StatusCode);
                 }
 
@@ -1206,10 +1226,12 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
 
                 input.Application.Output.Complete();
 
+#pragma warning disable CS0618 // Type or member is obsolete
                 var ex0 = Assert.Throws<BadHttpRequestException>(() => reader.TryRead(out var readResult));
                 var ex1 = Assert.Throws<BadHttpRequestException>(() => reader.TryRead(out var readResult));
                 var ex2 = await Assert.ThrowsAsync<BadHttpRequestException>(() => reader.ReadAsync().AsTask());
                 var ex3 = await Assert.ThrowsAsync<BadHttpRequestException>(() => reader.ReadAsync().AsTask());
+#pragma warning restore CS0618 // Type or member is obsolete
 
                 Assert.Equal(RequestRejectionReason.UnexpectedEndOfRequestContent, ex0.Reason);
                 Assert.Equal(RequestRejectionReason.UnexpectedEndOfRequestContent, ex1.Reason);
@@ -1231,10 +1253,12 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
 
                 input.Application.Output.Complete();
 
+#pragma warning disable CS0618 // Type or member is obsolete
                 var ex0 = Assert.Throws<BadHttpRequestException>(() => reader.TryRead(out var readResult));
                 var ex1 = Assert.Throws<BadHttpRequestException>(() => reader.TryRead(out var readResult));
                 var ex2 = await Assert.ThrowsAsync<BadHttpRequestException>(() => reader.ReadAsync().AsTask());
                 var ex3 = await Assert.ThrowsAsync<BadHttpRequestException>(() => reader.ReadAsync().AsTask());
+#pragma warning restore CS0618 // Type or member is obsolete
 
                 Assert.Equal(RequestRejectionReason.UnexpectedEndOfRequestContent, ex0.Reason);
                 Assert.Equal(RequestRejectionReason.UnexpectedEndOfRequestContent, ex1.Reason);
